Cleanup comment code
